fix typos

+++ remove_aux.py
@@ -6,12 +6,15 @@ to_remove = ['aux', 'gz', 'fls', 'log', 'out',
             'bcf', 'xml', 'dvi', 'pdf']
 
 if __name__ == "__main__":
-    for item in to_remove:            
-        aux_list = glob.glob(f"**/*.{item}", recursive = True)
-        
-        for dir in aux_list:    
-            try:
-                os.remove(dir)
-                print(f"Removed {dir}")
-            except:
-                print(f"Error removing {dir}")
+    if 'tex' in to_remove:
+        raise Exception("\nDo not remove .tex\n")
+    else:    
+        for item in to_remove:            
+            aux_list = glob.glob(f"**/*.{item}", recursive = True)
+            for dir in aux_list:    
+                try:
+                    os.remove(dir)
+                    print(f"Removed {dir}")
+                except:
+                    print(f"Error removing {dir}")
+        print('\nDone.')
